make total_distance integer
Former-commit-id: 8116e92cc85f86e48d8c9df8b3fee5f85935c2aa
Former-commit-id: 6b67edc447d79e66c2ca80239bbb65adfec6e567

+++ src/models/route.py
@@ -11,7 +11,7 @@ class GeoJsonGeometry(BaseModel):
 
 class GeoJsonProperty(BaseModel):
     sunlight_rate: float
-    total_distance: float
+    total_distance: int
     total_minutes: int
